Refactor CLI command orchestration
